Fetch providers list and pass them to client
